Move the healthcheck server to a shared module
