+++ market_depth.js
@@ -1,7 +1,7 @@
 const package = require('./package.json')
 const ccxt = require('ccxt')
 const { Exporter } = require('@santiment-network/san-exporter')
-const { marketDepth } = require('./utils')
+const { marketDepth, healthcheckServer } = require('./utils')
 const merge = require('lodash/merge')
 const exporter = new Exporter(`${package.name}-market-depth`)
 
@@ -43,26 +43,4 @@ async function main() {
 
 main()
 
-const healthcheckKafka = () => {
-  return new Promise((resolve, reject) => {
-    if (exporter.producer.isConnected()) {
-      resolve()
-    } else {
-      reject("Kafka client is not connected to any brokers")
-    }
-  })
-}
-
-module.exports = async (request, response) => {
-  const req = url.parse(request.url, true);
-
-  switch (req.pathname) {
-    case '/healthcheck':
-      return healthcheckKafka()
-        .then(() => send(response, 200, "ok"))
-        .catch((err) => send(response, 500, `Connection to kafka failed: ${err}`))
-
-    default:
-      return send(response, 404, 'Not found');
-  }
-}
+module.exports = healthcheckServer(exporter)
